Rc/v23.36.0

+++ platforms/docker/images/common-systemd/systemctl-commands.sh
@@ -13,7 +13,9 @@ fi
 
 if [[ $ID = "sles" ]]; then
     zypper -n install systemd
-    ln -s /usr/lib/systemd/systemd /sbin/init
+    if [ ! -e "/sbin/init"]; then
+        ln -s /usr/lib/systemd/systemd /sbin/init
+    fi
     zypper -n install awk
 fi
